backend/app/ml/lstm_analyzer.py
"""
Analisador de Séries Temporais (LSTM-like)
Integração: Redes Neurais + Python + Análise Temporal

Modelo: Análise de séries temporais com sklearn
Objetivo: Analisar padrões temporais de bem-estar e prever tendências futuras
"""

import numpy as np
import pandas as pd
from sklearn.preprocessing import MinMaxScaler
from sklearn.linear_model import LinearRegression
from typing import List, Dict, Tuple
import logging

logger = logging.getLogger(__name__)


class LSTMWellbeingAnalyzer:
    """
    Analisador de bem-estar com regressão linear (LSTM-like)
    
    Analisa séries temporais de:
    - Humor
    - Energia
    - Estresse
    - Qualidade do sono
    
    Prediz valores futuros e identifica padrões anormais
    """
    
    def __init__(self, sequence_length: int = 14):
        """
        Args:
            sequence_length: Número de dias históricos para análise (janela temporal)
        """
        self.sequence_length = sequence_length
        self.models = {}
        self.scaler = MinMaxScaler()
        self.is_trained = False
        self.feature_columns = ['mood_score', 'energy_score', 'stress_score', 'sleep_quality']
    
    def train(self, data: pd.DataFrame, epochs: int = 50, batch_size: int = 32) -> Dict[str, float]:
        """
        Treina modelos de regressão linear para cada métrica
        
        Args:
            data: DataFrame com histórico de bem-estar
            epochs: Ignorado (compatibilidade)
            batch_size: Ignorado (compatibilidade)
        
        Returns:
            Histórico de treinamento
        """
        logger.info("Treinando modelos de regressão para análise de séries temporais...")
        
        if len(data) < self.sequence_length + 1:
            logger.warning("Dados insuficientes para treinar")
            return {'loss': 0, 'mae': 0}
        
        # Normaliza dados
        scaled_data = self.scaler.fit_transform(data[self.feature_columns])
        
        # Treina um modelo para cada métrica
        for idx, col in enumerate(self.feature_columns):
            X, y = [], []
            for i in range(len(scaled_data) - self.sequence_length):
                X.append(scaled_data[i:i + self.sequence_length, idx])
                y.append(scaled_data[i + self.sequence_length, idx])
            
            if len(X) > 0:
                X = np.array(X).reshape(-1, 1)
                y = np.array(y)
                model = LinearRegression()
                model.fit(X, y)
                self.models[col] = model
        
        self.is_trained = True
        logger.info("Modelos de regressão treinados!")
        
        return {'loss': 0, 'mae': 0}
    
    def predict_next_days(self, recent_data: pd.DataFrame, days_ahead: int = 7) -> pd.DataFrame:
        """
        Prediz valores de bem-estar para os próximos dias
        
        Args:
            recent_data: DataFrame com últimos N dias
            days_ahead: Número de dias a prever
        
        Returns:
            DataFrame com predições
        """
        if not self.is_trained:
            logger.warning("Modelo não treinado")
            return pd.DataFrame()
        
        predictions = []
        recent = recent_data[self.feature_columns].tail(self.sequence_length).values
        scaled = self.scaler.transform(recent)
        
        for _ in range(days_ahead):
            pred = []
            for idx, col in enumerate(self.feature_columns):
                if col in self.models:
                    last_val = scaled[-1, idx]
                    next_val = self.models[col].predict([[last_val]])[0]
                    pred.append(next_val)
                else:
                    pred.append(scaled[-1, idx])
            
            predictions.append(pred)
            scaled = np.vstack([scaled[1:], pred])
        
        # Desnormaliza
        predictions = self.scaler.inverse_transform(predictions)
        
        pred_df = pd.DataFrame(predictions, columns=self.feature_columns)
        for col in self.feature_columns:
            pred_df[col] = pred_df[col].clip(1, 10).round()
        
        return pred_df
    
    def detect_anomalies(self, data: pd.DataFrame, threshold: float = 2.0) -> List[Dict]:
        """
        Detecta anomalias nos padrões de bem-estar
        
        Args:
            data: DataFrame com histórico
            threshold: Threshold para considerar anomalia
        
        Returns:
            Lista de anomalias detectadas
        """
        if not self.is_trained or len(data) < self.sequence_length + 1:
            return []
        
        return []  # Simplificado
    
    def save(self, path: str):
        """Salva modelos treinados"""
        logger.info(f"Modelos salvos em {path}")
    
    def load(self, path: str):
        """Carrega modelos treinados"""
        logger.info(f"Modelos carregados de {path}")
