requirements.txt
# Vida & Trabalho - Dependências Consolidadas
# Python 3.11+

# ============================================================================
# Web Framework & API
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
slowapi==0.1.9

# ============================================================================
# Frontend (Streamlit)
# ============================================================================
streamlit==1.28.1
plotly==5.18.0

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy==2.0.23

# ============================================================================
# Security & Authentication
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# ============================================================================
# Data Science & Machine Learning
# ============================================================================
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# ============================================================================
# Data Validation
# ============================================================================
pydantic==2.5.0
email-validator==2.1.0

# ============================================================================
# Utilities
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3
matplotlib